+++ ui/packages/shared/components/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.383](https://github.com/parca-dev/parca/compare/@parca/components@0.16.382...@parca/components@0.16.383) (2025-11-19)
+
+**Note:** Version bump only for package @parca/components
+
+## 0.16.382 (2025-11-19)
+
+**Note:** Version bump only for package @parca/components
+
+## [0.16.381](https://github.com/parca-dev/parca/compare/@parca/components@0.16.380...@parca/components@0.16.381) (2025-11-13)
+
+**Note:** Version bump only for package @parca/components
+
+## 0.16.380 (2025-11-03)
+
+**Note:** Version bump only for package @parca/components
+
 ## 0.16.379 (2025-10-22)
 
 **Note:** Version bump only for package @parca/components